backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
email-validator
alembic
passlib
bcrypt==3.2.2
PyJWT
python-multipart

celery
redis